Updated crawler to integrate with the sync-service
