--- nodefetch/src/api-request.ts
@@ -1,12 +1,0 @@
-export interface ApiRequest
-{
-    ttl: number;
-    source_uuid?: string;
-    name: string;
-    description: string;
-    website?: string;
-    phone?: string;
-    email?: string;
-    category?: string;
-    delivery?: string;
-}
